chore: move TestProcessWirePayForData
